- added delete for admin crud - reparse command: run parser on previously scraped job listings - made some column nullable, refactored reporter queries - fixed NoFluffJobsCrawler.php with urlencode - added seeder for crawlerkeywords

+++ laravel-backend/app/Services/Crawler/DTOs/Listing.php
@@ -8,8 +8,8 @@ class Listing
 {
     private string $position;
     private SalaryType $salaryType;
-    private float $salaryLow;
-    private float $salaryHigh;
+    private int $salaryLow;
+    private int $salaryHigh;
     private string $salaryCurrency;
     private int $homeOfficeDays = 0;
     private string $location;
